advanced plotting + docker + clean up INF
