+++ environment.yml
@@ -4,9 +4,9 @@ channels:
 dependencies:
   - python=3.12
   - openmm=8.2
+  - numpy=1.26.4
   - swig
   - pip
-  - numpy
   - numba
   - pdbfixer
   - biopython
