+++ scripts/OPES_analysis/analysis.py
@@ -8,8 +8,8 @@ import logging
 
 from src.analysis.fes import compute_fes
 from src.constants import kB
-import pandas as pd
-from src.analysis.deltaG import marginalize_fes
+from src.analysis.colvar import read_colvar_file
+from src.analysis.kernels import get_sigmas
     
 
 logger = logging.getLogger(__name__)
@@ -46,140 +46,28 @@ def consider_walls(df):
     df["total_bias"] = df["opes.bias"] + df["uwall.bias"]
     return df
 
-def get_sigma(directory, target, binder, run, cv):
-    # get the sigma from the kernels file
-    kernels_file = f"{directory}/{target}_{binder}.kernels"
-    df = pd.read_table(
-        kernels_file,
-        dtype=float,
-        sep=r"\s+",
-        comment="#",
-        header=None,
-        names=["time", "cmap", "d", "sigma_cmap", "sigma_d", "height", "logweight"]
-    )
-    sigma = df[f"sigma_{cv}"].iloc[-1]
-    return sigma.item()
+from src.analysis.plot import plot_all_fes_from_data
+def plot_all_fes(directory, target, binder, num_runs, labels):
+    """Load FES data and create plots
+    
+    Args:
+        directory: Base directory containing run folders
+        target: Target protein name 
+        binder: Binder name
+        num_runs: Number of runs to plot
+    """
+    # TODO: check, this might be broken now!!!!
+    fes_data = []
+    for run in range(1, num_runs + 1):
+        fes_filepath = f"{directory}/{binder}_{run}/{target}_{binder}_fes.h5py"
+        _, fes, cv1_bins, cv2_bins = load_fes(fes_filepath)
+        fes_data.append((cv1_bins, cv2_bins, fes))
 
+    plot_all_fes_from_data(fes_data, directory,target, binder, labels=labels)
 
-from mpl_toolkits.axes_grid1 import make_axes_locatable
-
-
-def plot_2d_fes(fes, cv1_bins, cv2_bins, cvs, ax):
-    cv1, cv2 = cvs
-
-    ax.set_xlabel(cv1)
-    ax.set_ylabel(cv2)
-
-    # Add filled contours
-    cntr = ax.contourf(cv1_bins, cv2_bins, fes, levels=range(0, 120, 5), cmap=plt.cm.jet)
-
-    # Add colorbar
-    divider = make_axes_locatable(ax)
-    cax = divider.append_axes("right", size="5%", pad="20%")
-    cbar = plt.colorbar(cntr, cax=cax, label="FES [kJ/mol]")
-    cbar.ax.yaxis.set_label_position("left")
-
-    # Add kBT scale
-    cbar_kBT = cbar.ax.twinx()
-    cbar_kBT.set_ylim(cbar.vmin / (kB * TEMP), cbar.vmax / (kB * TEMP))
-    cbar_kBT.set_ylabel("FES [kBT]", rotation=270, labelpad=15)
-
-    return ax
-
-def get_sigmas(directory, target, binder, run, cvs):
-    assert len(cvs) == 2, "Only 2D FES are supported"
-    return [get_sigma(directory, target, binder, run, cv) for cv in cvs]
-
+from src.analysis.plot import plot_2d_fes, plot_1d_fes
+from src.analysis.fes import load_fes
 # compute FES
-def read_colvar_file(filename):
-    # Read first line to get column names
-    with open(filename) as f:
-        header_line = f.readline().strip()
-    
-    # Parse column names from FIELDS line
-    if not header_line.startswith('#! FIELDS'):
-        raise ValueError("First line must start with '#! FIELDS'")
-    column_names = header_line.replace('#! FIELDS', '').strip().split()
-    
-    # Read data using column names from file
-    colvar_df = pd.read_table(
-        filename,
-        dtype=float,
-        sep=r"\s+",
-        comment="#", 
-        header=None,
-        names=column_names
-    )
-
-    # remove the unbiased bit
-    # PLUMED:   adaptive SIGMA will be used, with ADAPTIVE_SIGMA_STRIDE = 5000
-    # PLUMED:     thus the first x kernel depositions will be skipped, x = ADAPTIVE_SIGMA_STRIDE/PACE = 10
-    # get the pace from the _plumed.dat file
-    filename = filename.replace(".colvar", "_plumed.dat")
-    with open(filename, "r") as f:
-        content = f.read()
-    
-    # depositing kernels
-    pace_pattern = r"opes:\s*OPES_METAD[\s\S]*?PACE=(\d+)"
-    match = re.search(pace_pattern, content)
-    if match:
-        pace = int(match.group(1))
-        logger.info(f"Pace value: {pace}")
-    else:
-        raise ValueError("PACE not found in the file")
-    
-    # printing to COLVAR file
-    stride_pattern = r"PRINT.*STRIDE=(\d+)"
-    match = re.search(stride_pattern, content)
-    if match:
-        stride = int(match.group(1))
-        logger.info(f"Stride value: {stride}")
-    else:
-        raise ValueError("Stride not found in the file")
-    
-    # assume first x = 10 kernel depositions are the ones to skip
-    # compute how many printing steps correspond to that
-    num_steps = pace * 10
-    strides_to_skip = num_steps // stride
-    logger.info(f"Skipping {strides_to_skip} steps")
-
-    # look at the opes.bias in the strides_to_skip steps and assert they are constant
-    assert colvar_df["opes.bias"].iloc[:strides_to_skip].std() < 1e-6, "opes.bias is not constant in non-biased part"
-    colvar_df = colvar_df.iloc[strides_to_skip:]
-    return colvar_df
-
-def plot_1d_fes(fes, cv1_bins, cv2_bins, cvs, axs):
-    ax1, ax2 = axs
-    cv1, cv2 = cvs
-    # plot along first CV
-    # axis = 0, as we keep the first CV
-    fes_1d = marginalize_fes(fes, kBT=kB*TEMP, axis=0)
-    ax1.plot(cv1_bins, fes_1d)
-    ax1.set_xlabel(cv1)
-    ax1.set_ylabel("FES [kJ/mol]")
-
-    # plot along second CV
-    fes_1d = marginalize_fes(fes, kBT=kB*TEMP, axis=1)
-    ax2.plot(cv2_bins, fes_1d)
-    ax2.set_xlabel(cv2)
-    ax2.set_ylabel("FES [kJ/mol]")
-    ax2.invert_xaxis()
-    return ax1, ax2
-
-def load_fes(filepath: str):
-    with h5py.File(filepath, "r") as f:
-        cvs = f.attrs['cvs']
-        fes = f["fes"][:]
-        cv1_bins = f[f"{cvs[0]}_bins"][:]
-        cv2_bins = f[f"{cvs[1]}_bins"][:]
-        first_axis = f.attrs['1st_axis']
-        second_axis = f.attrs['2nd_axis']
-
-    assert first_axis in cvs and second_axis in cvs, "First and second axis must be in CVs"
-    assert first_axis != second_axis, "First and second axis must be different"
-    assert first_axis == cvs[0] and second_axis == cvs[1], "First axis must be the first CV and second axis must be the second CV"
-    assert fes.shape[0] == len(cv1_bins) and fes.shape[1] == len(cv2_bins), "FES must have the same shape as the CV bins"
-    return cvs, fes, cv1_bins, cv2_bins
 
 def plot_bias(colvar_df, ax):
     ax.plot(colvar_df["time"] * TIMESTEP * OPES_PACE, colvar_df["opes.bias"] + colvar_df["uwall.bias"], label='total')
@@ -242,82 +130,6 @@ def plot_everything(directory, target, binder, run):
         edgecolor='none'
     )
     plt.close()
-
-
-def plot_all_fes(directory, target, binder, num_runs):
-    # plot all the FES for a given system in a single plot
-    # make it scalable so that you have the same number of columns as runs
-    # first row: full FES
-    # second row: 1D FES along first CV
-    # third row: 1D FES along second CV
-
-    # Calculate figure width based on number of runs to maintain square subplots
-    # Height is fixed at 15 inches, width scales with number of runs
-    fig_height = 15
-    subplot_size = fig_height / 3  # Each subplot should be square
-    fig_width = subplot_size * num_runs * 1.2  # Add 20% more width for spacing
-    
-    # Create figure with square-looking subplots
-    fig = plt.figure(figsize=(fig_width, fig_height))
-    fig.suptitle(f"{binder=} and {target=}", fontsize=16, y=0.98)
-
-    # Use gridspec with adjusted spacing
-    gs = fig.add_gridspec(3, num_runs, wspace=0.3, hspace=0.3, top=0.95)
-    axs = np.empty((3, num_runs), dtype=object)
-    
-    # Create subplots using gridspec
-    for i in range(3):
-        for j in range(num_runs):
-            axs[i,j] = fig.add_subplot(gs[i,j])
-            if i == 0:  # Add run number at the top of each column
-                axs[i,j].set_title(f"run={j+1}", pad=5)  # Reduced pad value
-
-    cv1_mins = []
-    cv1_maxs = []
-    cv2_mins = []
-    cv2_maxs = []
-
-    for run in range(1, num_runs + 1):
-        fes_filepath = f"{directory}/{binder}_{run}/{target}_{binder}_fes.h5py"
-        cvs, fes, cv1_bins, cv2_bins = load_fes(fes_filepath)
-        axs[0, run - 1] = plot_2d_fes(fes, cv1_bins, cv2_bins, cvs, axs[0, run - 1])
-        axs[1, run - 1], axs[2, run - 1] = plot_1d_fes(fes, cv1_bins, cv2_bins, cvs, axs[1:3, run - 1])
-
-        cv1_mins.append(cv1_bins[0])
-        cv1_maxs.append(cv1_bins[-1])
-        cv2_mins.append(cv2_bins[0])
-        cv2_maxs.append(cv2_bins[-1])
-
-    cv1_min = min(cv1_mins)
-    cv1_max = max(cv1_maxs)
-    cv2_min = min(cv2_mins)
-    cv2_max = max(cv2_maxs)
-
-    # set limits on all subplots
-    # first row, use the cv1 and so for each of the axis
-    for j in range(num_runs):
-        axs[0, j].set_xlim(cv1_min, cv1_max)
-        axs[0, j].set_ylim(cv2_min, cv2_max)
-    # second row, change based on cv1
-    for j in range(num_runs):
-        axs[1, j].set_xlim(cv1_min, cv1_max)
-    # third row, change based on cv2
-    for j in range(num_runs):
-        # this one is flipped
-        axs[2, j].set_xlim(cv2_max, cv2_min)
-    
-
-
-    plt.savefig(
-        f"{directory}/{binder}_all_fes.png",
-        dpi=300,
-        bbox_inches='tight',
-        facecolor='white',
-        edgecolor='none'
-    )
-    plt.close()
-
-    return    
 
 from matplotlib.animation import FuncAnimation, PillowWriter
 from tqdm import tqdm
@@ -456,8 +268,7 @@ def run(date, systems, num_runs, recompute, collect_plots):
                 
             plot_colvar_traj_in_fes(system_directory, target, binder, num_runs)
             shutil.copy(f"{system_directory}/{target}_{binder}_all_trajectories.gif", f"{save_dir}/{target}_{binder}_all_trajectories.gif")
-
-            plot_all_fes(system_directory, target, binder, num_runs)
+            plot_all_fes(system_directory, target, binder, num_runs, labels=[f"{run=}" for run in range(1, num_runs + 1)])
             shutil.copy(f"{system_directory}/{binder}_all_fes.png", f"{save_dir}/{target}_{binder}_all_fes.png")
